RecommdationFinish

+++ src/main/java/com/example/mbti/repository/UserRepository.java
@@ -1,9 +1,13 @@
 package com.example.mbti.repository;
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.mbti.model.User; 
+import com.example.mbti.model.User;
+ 
 
 
 
-public interface UserRepository  extends JpaRepository<User, Long> {
+public interface UserRepository  extends JpaRepository<User, Integer> {
     User findByEmail(String email); // for login
+    //Optional findById(Integer id);
 }
